Move ChatRoom[Impl] to org.jitsi.jicofo.xmpp.muc.
