+++ jicofo/src/main/kotlin/org/jitsi/jicofo/conference/ConferenceUtil.kt
@@ -17,9 +17,9 @@
  */
 package org.jitsi.jicofo.conference
 
-import org.jitsi.impl.protocol.xmpp.ChatRoom
 import org.jitsi.jicofo.visitors.VisitorsConfig
 import org.jitsi.jicofo.xmpp.XmppProvider
+import org.jitsi.jicofo.xmpp.muc.ChatRoom
 import org.jitsi.utils.MediaType
 import org.jitsi.xmpp.extensions.colibri2.Media
 import org.jitsi.xmpp.extensions.jingle.ContentPacketExtension
